requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coverage==5.3
distlib==0.3.0
distro==1.4.0
flake8==3.8.4
Flask==1.1.2
Flask-Testing==0.8.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
packaging==20.3
pep517==0.8.2
progress==1.5
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pyparsing==2.4.6
python-dotenv==0.15.0
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
snowballstemmer==2.0.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
